algorithms: math for calculating portions of "scoring"
Multiple different functions for calculating a score. While most of this could just be client-side since we have to accept their data input anyway, I'd rather be able to limit the attack surface area given for cheaters.

For example: calculating accuracy means fully trusting player input for now since we don't really have a way to get the maximum notes on any given track, so that is assembled from the given input. Grades are also very dependent on the score that player received, there is currently no way we can entirely trust this input unless there is some form of anti-cheat validation on the other end, which I'm currently uninterested in. This is the same for medals and volforce, however volforce is special since it is specifically used to influence a score's ranking order when viewed ordinally, so not having the client handle this for us removes a lot of cases of very, very, lazy cheaters.