database: profile type struct and allow searching for them by uuid
Breaking microservice conventions again because I don't know a better way to do this. Wouldn't be needed if we had stored User's UUID's instead of IDs, or if JWT claims had the User's ID in them, but sadly we don't. So I have to resort to using this hack because I need to find the user's ID when they submit a score.