database: improve semantics on score submissions
Where a user has previously submitted a score to a board, we will check the performance rating. If the performance rating is higher than last time, then it will update the score instead of adding a new entry. If it is worse, we will discard the score, rewards should still be calculated, but that isn't handled here, so we will return a special error to signal to perform such tasks.